+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Over 1,000 marker definitions from four whole-genome screens from Yu et al 2022 (see #115, #118).
 - New `nomenclature` module for parsing and validating microhap identifiers (see #119).
 - New `summarize` subcommand for summarizing MicroHapDB database contents (see #120).
+- 500 marker definitions from Zhu et al 2023 (see #125).
 
 ### Changed
 - Overhaul of the marker table structure and database build procedure (see #106, #111, #112).